[FEAT] : server fin
