+++ apps/ServiceManager/src/battery-src/battery/battery.cpp
