+++ apps/ServiceManager/src/battery-src/battery_client/client-main.cpp
