+++ apps/ServiceManager/src/etc/service-example.hpp
@@ -11,7 +11,7 @@
 #include <thread>
 #include <mutex>
 #include <csignal>
-#include "sample-ids.hpp"
+#include "server.hpp"
 // #include "canreceiver.hpp"
 
 #define SPEED 1
