+++ apps/ServiceManager/src/gear-src/gear_data_receiving_client/client-example.hpp
@@ -9,7 +9,7 @@
 #include <string>
 #include <thread>
 #include <csignal>
-#include "server.hpp"
+#include "../../server.hpp"
 
 class client_sample {
 public:
