+++ apps/ServiceManager/src/gear-src/gear_data_receiving_client/client-main.cpp
