+++ apps/ServiceManager/src/headers.hpp
@@ -1,0 +1,25 @@
+#ifndef HEADERS_HPP
+# define HEADERS_HPP
+
+#include <vsomeip/vsomeip.hpp>
+#include <chrono>
+#include <condition_variable>
+#include <iomanip>
+#include <iostream>
+#include <sstream>
+#include <thread>
+#include <mutex>
+#include <csignal>
+#include <unistd.h>
+#include <sys/wait.h>
+#include <cstring>
+#include <sys/ioctl.h>
+#include <net/if.h>
+#include <linux/can.h>
+#include <linux/can/raw.h>
+#include <cstdint>
+#include <string>
+#include <memory>
+
+
+#endif
