+++ apps/ServiceManager/src/speed-src/speed_client/client-main.cpp
@@ -1,0 +1,35 @@
+#include "client-example.hpp"
+
+#ifndef VSOMEIP_ENABLE_SIGNAL_HANDLING
+client_sample *its_sample_ptr = nullptr;
+
+void handle_signal(int _signal) {
+    if (its_sample_ptr != nullptr &&
+        (_signal == SIGINT || _signal == SIGTERM)) {
+        its_sample_ptr->stop();
+    }
+}
+#endif
+
+int main(int argc, char **argv) {
+    bool use_tcp = false;
+
+    client_sample its_sample(use_tcp);
+
+
+#ifndef VSOMEIP_ENABLE_SIGNAL_HANDLING
+    its_sample_ptr = &its_sample;
+    signal(SIGINT, handle_signal);
+    signal(SIGTERM, handle_signal);
+#endif
+
+    if (its_sample.init()) {
+        its_sample.start();
+#ifdef VSOMEIP_ENABLE_SIGNAL_HANDLING
+        its_sample.stop();
+#endif
+        return 0;
+    } else {
+        return 1;
+    }
+}
